🚨(backend) various review fixes
I am doing various small review fixies
