--- src/backend/conversations/configuration/llm/dev.json
@@ -1,43 +1,0 @@
-{
-  "models": [
-    {
-      "hrid": "default-model",
-      "model_name": "settings.AI_MODEL",
-      "human_readable_name": "Default Model",
-      "provider_name": "default-provider",
-      "profile": null,
-      "settings": {},
-      "is_active": true,
-      "icon": [
-        "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABwAAAAcCAMAAABF0y+mAAAAn1BMVEUALosAKoovTZjw8vb////+9/jlPUniAAz",
-        "iABUAGIWbpsTwq7HhAAAAI4dle7DrdX4AJohRaaboXWj7+/zn6On5//9NZaT29vfoWmVHYKDoUl/k5OUAIYddc6vpbHYCM47Y3+v53+LiFCUA",
-        "HIWnsckYPJHi6PL77O7jJjW3wdf1w8jre4QgQ5TZ2txwg7Pr3+I8WZ6OnsTuoamClL7tlZ5xz5y8AAAAzUlEQVR4AZ3RRQKDQBBEUSTu7h5c4",
-        "vc/W6Yp3KG2Dz4ynDdeEBvOmq12xx2E1u0B+4NOEocj4DgNJ1PgLAvni8WyBq5Yc71ubFJx23C2q4P7dRYejg1xzvCUgvz5guz11k7gXYKF/1",
-        "8oyiYuvHAYeVkhXCzolVStHcGDjiQzNmMQxsMI5rEJRdQSPZvbpE2E8aY6gC6Z+2Hg4dFA0Yb4YedNL/v4Fk8WJuwiGhrChJNXI210rnib9Fs",
-        "JlXRUC/HwTscPIXf/iklq/tjb/gHAdxkCUjAg2QAAAABJRU5ErkJggg=="
-      ],
-      "system_prompt": "settings.AI_AGENT_INSTRUCTIONS",
-      "tools": "settings.AI_AGENT_TOOLS"
-    },
-    {
-      "hrid": "default-summarization-model",
-      "model_name": "settings.AI_MODEL",
-      "human_readable_name": "Default Summarization Model",
-      "provider_name": "default-provider",
-      "profile": null,
-      "settings": {},
-      "is_active": true,
-      "icon": null,
-      "system_prompt": "settings.SUMMARIZATION_SYSTEM_PROMPT",
-      "tools": []
-    }
-  ],
-  "providers": [
-    {
-      "hrid": "default-provider",
-      "base_url": "settings.AI_BASE_URL",
-      "api_key": "settings.AI_API_KEY",
-      "kind": "mistral"
-    }
-  ]
-}
